requirements.txt
apache-airflow==3.0.1
confluent_kafka==2.10.0
kafka_python==2.2.9
minio==7.2.15
pyspark==3.5.1
python-dotenv==1.1.0
Requests==2.32.3
